+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to `Console` will be documented in this file.
 
+## [0.2.0](https://github.com/staticka/staticka/compare/v0.1.1...v0.2.0) - Unreleased
+
+### Added
+
+- `Create` command
+- `Watch` command
+- Configuration (through `staticka` property)
+
+### Changed
+
+- `Build` command
+
+### Deleted
+
+- Options for `Build` command
+
 ## [0.1.1](https://github.com/staticka/staticka/compare/v0.1.0...v0.1.1) - 2018-04-08
 
 ### Changed